Refactored AirbrakeConfig into AirbrakeClientBuilder
